requirements.txt
pyyaml>=3.0,<4.0
elasticsearch>=5.0.0,<6.0.0
